+++ grid/social-media/upload-url/maps/instagram.suma
@@ -1,9 +1,9 @@
 // Instagram Graph API Reference: https://developers.facebook.com/docs/instagram-api
 
-profile = "social-media/upload-assets@1.0"
+profile = "social-media/upload-url@1.0"
 provider = "instagram"
 
-map UploadAssetViaUrl {
+map RegisterUpload {
   apiVersion = 'v14.0'
   businessAccountId = input.profileId
   authorizationHeaderValue = 'Bearer ' + parameters.accessToken
@@ -13,11 +13,11 @@ map UploadAssetViaUrl {
   }
   caption = input.caption
 
-  call UploadMedia(businessAccountId = businessAccountId, item = item, isCarousel = false, caption = caption, authorization = authorizationHeaderValue) {
+  call UploadMedia(businessAccountId = businessAccountId, item = item, caption = caption, type = input.type, authorization = authorizationHeaderValue) {
     return map result if (!outcome.error) {
       assetId = outcome.data
     }
-    return map error error
+    return map error outcome.error
   }
 }
 
@@ -91,18 +91,28 @@ operation UploadMedia {
   apiVersion = 'v14.0'
   businessAccountId = args.businessAccountId
   itemUrl = args.item ? args.item.url : undefined
-  isCarousel = args.isCarousel || undefined
   children = args.children ? JSON.stringify(args.children) : undefined
-  caption = isCarousel ? undefined : args.caption // captions are not supported for carousel items
+  caption = args.caption
   authorizationHeaderValue = args.authorization
+
+  mediaType = (() => {
+    switch (args.type) {
+      case 'video':
+        return 'VIDEO';
+      case 'reel':
+        return 'REELS';
+      default:
+        return undefined;
+    }
+  })()
 
   http POST "/{apiVersion}/{businessAccountId}/media" {
     request {
       query {
+        image_url = itemUrl,
         video_url = itemUrl,
         caption = caption,
-        is_carousel_item = isCarousel,
-        media_type = 'VIDEO',
+        media_type = mediaType,
         children = children,
       }
