+++ grid/social-media/upload-url/maps/instagram.test.ts
@@ -1,6 +1,7 @@
-import { publishVideoTest } from './upload-assets';
+import { publishVideoTest } from './upload-url';
 
-publishVideoTest('instagram', {
+publishVideoTest('instagram', 'single video', {
   caption: 'Single video from Superface Station.',
   url: 'https://download.samplelib.com/mp4/sample-5s.mp4',
+  type: 'video',
 });
