+++ grid/social-media/upload-url/maps/mock.suma
@@ -1,0 +1,14 @@
+profile = "social-media/upload-url@1.0"
+provider = "mock"
+
+map RegisterUpload {
+  return map result {
+    assetId = '125682506616099'
+  }
+}
+
+map GetAssetState {
+  return map result {
+      state = 'Finished'
+  }
+}
