+++ grid/social-media/upload-url/maps/mock.test.ts
@@ -1,0 +1,6 @@
+import { publishVideoTest } from './upload-url';
+
+publishVideoTest('mock', 'single video', {
+  caption: 'Single video from Superface Station.',
+  url: 'https://download.samplelib.com/mp4/sample-5s.mp4',
+});
