+++ superface/super.json
@@ -445,11 +445,14 @@
         }
       }
     },
-    "social-media/upload-assets": {
-      "file": "../grid/social-media/upload-assets/profile.supr",
+    "social-media/upload-url": {
+      "file": "../grid/social-media/upload-url/profile.supr",
       "providers": {
+        "mock": {
+          "file": "../grid/social-media/upload-url/maps/mock.suma"
+        },
         "instagram": {
-          "file": "../grid/social-media/upload-assets/maps/instagram.suma"
+          "file": "../grid/social-media/upload-url/maps/instagram.suma"
         }
       }
     },
